GEOMESA-876 Move csv package from accumulo-ds
Moved the csv package from accumulo-datastore to geomesa-utils.
