+++ geomesa-utils/pom.xml
@@ -36,6 +36,10 @@
             <artifactId>commons-pool2</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-csv</artifactId>
+        </dependency>
+        <dependency>
             <groupId>commons-codec</groupId>
             <artifactId>commons-codec</artifactId>
         </dependency>
@@ -82,6 +86,10 @@
         <dependency>
             <groupId>org.geotools</groupId>
             <artifactId>gt-data</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.geotools</groupId>
+            <artifactId>gt-shapefile</artifactId>
         </dependency>
         <dependency>
             <groupId>joda-time</groupId>
