+++ geomesa-utils/src/main/scala/org/locationtech/geomesa/utils/csv/DMS.scala
@@ -6,7 +6,7 @@
 * http://www.opensource.org/licenses/apache2.0.php.
 *************************************************************************/
 
-package org.locationtech.geomesa.accumulo.csv
+package org.locationtech.geomesa.utils.csv
 
 // TODO: replace with org.geotools.measure.AngleFormat
 object DMS {
@@ -45,7 +45,7 @@ object DMS {
   }
 }
 
-import org.locationtech.geomesa.accumulo.csv.DMS._
+import org.locationtech.geomesa.utils.csv.DMS._
 
 case class DMS(degrees: Int, minutes: Int, seconds: Double, hemisphere: Hemisphere) {
   require(0 <= seconds && seconds < 60, "Seconds must be between 0 and 60")
